[update] some optimization
